requirements.txt
# ADB 安卓截图 + 离线 OCR 识别脚本依赖
# 注意：PaddlePaddle 3.3.x 在 Windows CPU 环境下存在 oneDNN 兼容性 bug，
#       必须固定使用 3.2.2 版本。

# 1. PaddlePaddle 基础推理库（CPU 版）
# 国内用户可使用清华镜像加速：-i https://pypi.tuna.tsinghua.edu.cn/simple
paddlepaddle==3.2.2

# 2. PaddleOCR 离线 OCR 库
paddleocr>=2.7.0

# 3. 图像处理
Pillow>=10.0.0

# 4. Excel 导出
openpyxl>=3.1.0

# 5. 其他工具（通常已内置，显式声明避免遗漏）
numpy>=1.24.0
